refactor: Expression DataFrame Preprocessing Update
The previous implementation of process expression compendium did not properly handle non-numeric columns for sample names and did not concatenate each expression compendium correctly.

The new implementation also allows the variance threshold and the minimum expression filtering to be passed as arguments instead of having constants defined in the preprocessing.py file. In addition, the new implementation uses only numpy and does not do extra conversions of data frames to use scanpy.