feature: legend pick event callback for toggle selection
Toggling point selection helps visualize points that are getting covered or compare a few data sets out of the larger compendium at a time.

This implementation defines a callback for matploltlibs pick_event. Each data sets legend label is set to fire a pick_event when it's clicked. Each legend item also has a mapped scatter set so the on_legend_click can hide/show the points for whatever legend option is clicked.

The generate layout script defines TkAgg as the interactive backend for matplotlib so that event handling is included.

Note: seaborn.scatterplot() returns an AxesSubplot object and not a PathCollection. We need to use a PathCollection to adjust point properties. Replace creating the scatter plots for each compendium using seaborn with using matplotlib base.