refactor: plot generation into plotting module in src
Plotting code does not fit the purpose of a layout algorithm. All types of layout algorithms should preform some mathematical calculation on the expression df and return a standard format. This standard format should be handed off to some other level of abstraction for plotting, in this case the plotting module.