feat: add_scatter_plots custom unique_labels param and plot diseases
Scatter plots by default will be created for every unique value of a label. If the caller chooses, unique_labels can be predfined to a specific list of values. This allows labels that have a lot of unique values to specify which values they want plotted.

Define a generate_disease_plot that works the same as generating a compendium plot but defines a custom unique_labels list for the 9 diseases with the most data and the points that belong to unkown diseases.