feat: customize figure file name using gen_figure_file_path in config
Support for generating figure file names from within a script. the gen_figure_file_path accepts a file name argument and returns a path to where figure file can be downloaded. This is useful for scripts that change the number of figures they save or want to save custom names for different labels for example.